stbt camera: Use inverse homography matrix
Rather than the homography matrix.  The inverse is the one which we
have to use in the transformation.  This will make porting to OpenGL
easier.

The identity inverse homography matrix can more succinctly be exactly
represented as rather than it being approximately

      0.6666666   0.0         -0.1666666
    [ 0.0         0.6666666   -0.1666666 ]
      0.0         0.0          1.0

the identity inverse homography matrix is exactly

      1.5  0.0  0.25
    [ 0.0  1.5  0.25 ]
      0.0  0.0  1.0

This was causing better matches in the test
`test_that_validation_passes_on_pristine_input` tripping over a bug in
`stbt camera validate` indexing into the string `"0123456789"` with the
value `10` because it was a perfect match.