FrameObject: Make FrameObject a metaclass
With this change the list of attributes is created at class definition time
and attached to the class, rather than being created at object creation
time and being attached to each instance.

This will be important for adding caching and changning properties to be
`None` if `is_visible` is `False`.