stbt auto-selftest: Improve isolation with multiprocessing
stbt auto-selftest imports a bunch of Python code of unknown quality.
With this commit we only do this in subprocesses where nothing untoward
can happen.

This isn't in response to any specific issue I've had, but it's easy to
do and just seems like a good idea.  Perhaps we can use this technique in
the future to do more in parallel.