stbt rig: Add git_snapshot function
Takes a snapshot of your git repo working directory creating a new commit
leaving your git repo otherwise unaffected.  Returns the commit SHA.

The commit will have the message "snapshot".

This function will raise UnknownFileInRepo exception if there are files that
have not been added nor ignored in the repo.

If there are no changes in the working directory it will just return the
commit SHA of HEAD.

We will use this function to implement a `stbt rig sync` which will keep
a remote branch up-to-date with your working directory.