_stbt.core: Remove GObject.threads_init for ancient versions of GObject
`threads_init` is not needed for PyGObject 3.10.2+ (and it started
printing a deprecation warning from 3.12.0).

There is no supported version of Fedora or Ubuntu with a version as old
as that:

* Fedora 22 has 3.16.2
* Ubuntu 14.04 has 3.12.0
* stb-tester doesn't support Ubuntu 12.04 as it doesn't have
  GStreamer 1.x.

I'm removing this to make the GObject/Gst initialisation in core.py the
same as in gst_utils.py.