_stbt: Refactor image debug logging to use ImageLogger class
Instead of writing debug images to disk directly, `match`,
`detect_motion` and `is_screen_black` register the debug images with an
instance of `ImageLogger`. Helper functions like `_find_match` are
injected the `ImageLogger` instance (this is the observer pattern).

This is better because previously `_log_image` (which wrote images to
disk) and `_log_image_descriptions` (which wrote an html file)
communicated with each other via the filesystem --
`_log_image_descriptions` would list the files on disk to determine how
many levels of pyramid-matching had happened. This would create
incorrect html output if a previous invocation of `stbt run` had created
3 levels of debug screenshots in `stbt-debug/detect_match/00001/` but
the current invocation had only created 1 level (because no match was
found).

It's also less magical -- previously we incremented the frame counter
when someone logged an image called "source.png". Now we increment the
frame counter when you create a new instance of `ImageLogger`.

My motivation in doing this work is that I want to add additional
logging for https://github.com/stb-tester/stb-tester/pull/212 so I
want to simplify the code first.