Caching: Add xxhash library submodule at v0.5.1
We will be using this to perform hashing of uncompressed video frames.  In
measurements on my machine it can hash a 720p BGR image in ~240us, whereas
`hash()` and `hashlib.sha1` both take >3ms.

Also, unlike the `hash()` builtin the results should be sharable between
Python versions and computer architectures, so it's far more appropriate
to store it in a file.

64-bits should be fine as a hash size.  With 6 million hash values there
should only be 1 in a million chance of a collision[1].

[1]: http://preshing.com/20110504/hash-collision-probabilities/