Refactor _match logging so it's more ameniable for caching
Previously a `template_name` parameter was passed into _match.  This was
only used to provide additional information for writing out the debug
html.  This parameter contained the full path to the template which was
defeating caching in testing because we take a hash of all the input
arguments for caching.

This commit refactors this debug html writeout by moving some of it into
the parent `match()` function.  By doing this we can pass the
`MatchResult` into `_log_match_image_debug`.  This is better as the
`MatchResult` is user facing, so it's clearer which implementation details
of `_match` we're exposing that aren't a part of the public API.