Python API: Add stbt.match_all
Returns a list of `MatchResult` objects, so that you can find all
instances of a particular image that are on the screen at the same time.

Based on an earlier patch by William Manley.

The debug image logging (`stbt match -v` and `stbt run -vv`) still logs
a single match (probably the last one seen). This will be addressed in a
future commit.

This removes caching from `stbt.match` (because `_find_matches` is now a
generator, and `imgproc_cache.memoize` doesn't know how to cache the
results from a generator). This will be added back in a future commit.