tests: Add test for the region reported for a failed match
If the match fails at the first level of the pyramid algorithm (when
we're searching in a down-scaled version of the frame) then we have to
up-scale the region we report for the best (but not good enough) match.

This passes currently, but there was no test for it and I want to make
sure I don't break it when I implement `stbt.match_all`.