auto-selftest: Add caching of expensive image processing functions
This dramatically speeds up re-running `stbt auto-selftest generate` when
you have a lot of FrameObjects × a lot of selftest screenshots.

On a private repo with 196 selftest-screenshots and 73 expressions
(14308 doctest items to evaluate) the results are rather promising:

* 1:01:13 without caching, median expression time 120ms
* 37:50 with a cold cache, median expression time 38ms (1.6x speed up)
* 2:20 total with a hot cache. Median time per expression: 3.4ms (26x
  speed up)

This generated an 8.2MB cache file with 22843 entries (361B/entry).

This makes stbt auto-selftest usable on large test-packs where it just
wasn't fast enough before to be usable.

We use the xxhash library for fast (non-crypto) image hashing. I evaluated
sha1 and python's hash() builtin and they were too slow, taking >3ms for
a 720p image. xxhash does the same in 240μs.

We use lmdb as the on-disk cache database because it's small, fast and has
good Python bindings.