auto-selftest: Nicer Ctrl-C behaviour
When you Ctrl-C `stbt auto-selftest` it can take a little while to shut
down as it waits for the file it's processing to complete.  This is due to
python bug [8296].  We work around this bug here by using a `apply_async`
and a timeout.

Also: previously we would see an error message printed from the child
process on Ctrl-C.  This was misleading as it didn't seem to help the
program to shut-down.  Now we block SIGINT in the child processes to avoid
this.

[8296]: http://bugs.python.org/issue8296