Caching: Add py-lmdb submodule v0.86
We will be using LMDB as backend storage for our image-processsing cache.
Unfortunately python-lmdb package is not available in Ubuntu 14.04, so
this commit adds it to our sources such that it will be bundled with
stb-tester.  This can be dropped once we drop support for Ubuntu 14.04.

v0.86 (2015-06-08) was chosen over the latest v0.89 (2016-02-12) as that's
what's currently available in Debian and is what was shipped with Ubuntu
16.04.