Rename "stbt templatematch" command to "stbt match"
For consistency with the Python API "stbt.match". I've been trying to
remove the wording "templatematch" from our public APIs as it is the
name of the OpenCV function we use in the first pass of our matching
algorithm; as such it is an implementation detail. The argument names to
`stbt.match`, `stbt.wait_for_match` etc are "image" (not "template") and
in prose we call this a "reference image" rather than a "template".

For backwards compatibility we keep "stbt templatematch" as an alias.