_stbt: Implement debug image logging for match_all
The debug output you get in `stbt-debug/` when you run `stbt run -vv` or
`stbt match -v` now shows all matches returned by `match_all`, not just
the first one.

I changed the html output to display the matches in a table instead of a
bullet list -- the table is better for displaying the multiple matches.