stbt-docker: Make sure we're actually setting STBT_CONFIG_FILE
Without this it gets replaced with STBT_CONFIG_FILE set in the test-pack
entrypoint.  In this instance it would just be overriden by the exact same
value so this could just be removed but we're going to have to start
setting this on Windows.  This is because git on Windows checks symlinks
out as normal files, which messes up any .stbt.conf indirection.