FrameObject: Anticipate stbt.get_frame returning a read-only image
In the future I want to make `stbt.get_frame()` return an immutable image.
This way we will be able to reduce copying of data by providing numpy
arrays backed by the original `Gst.Sample`s.  This will be particularly
important for better and more efficient handling of 4K video.

This commit anticipates that change by exercising user's `FrameObject`s
with read-only buffers.  This is good because it makes a future change to
the behaviour of `get_frame()` less likely to break user-scripts.  This is
a non-breaking change itself as `FrameObject` has yet to be released.