MatchParameter.frame is now read-only
This affects the MatchParameter object returned from `match`,
`detect_match`, `wait_for_match`, and the new `match_all`. It is a
breaking change but it probably won't affect anyone -- even if you *are*
doing anything with the frame, most of OpenCV's functions are
non-destructive.

It's still a copy, because `numpy_from_sample` unmaps it as soon as the
contextmanager exits. This is something that we hope to address in the
future; this change paves the way for that optimization in the future.

In the immediate term, this allows us to avoid an extra copy in each
iteration of `match_all`.