Attach a real-time capture timestamp to each GstSample received
The timestamps we currently provide are the PTS data attached to the
`Gst.Sample`s pulled from the source pipeline.  In practice with v4l2src
this will have tended to roughly be the time from the start of the test or
the last pipeline restart.  It could be some other number depending on the
source pipeline used.   It is only useful for comparing the time
difference between two frames.

This commit introduces `sample.time` for each pulled frame.
This is the time since the unix epoch in seconds as a float.  This can be
used to compare the time of a frame against:

* The current time
* Other events that happen on the system such as IR presses
* Other events that occur in the world such as remote server timestamps or
  for measuring synchronisation between different devices-under-test
  (assuming a correctly set system-clock).