HDMI CEC Control: Add build scaffolding in preparation for libcec
libcec is GPLv2 licensed, unlike the rest of stb-tester which is licenced
under the LGPLv2+.  This commit adds an additional package: stb-tester-gpl
which will be used to quarantine GPL dependencies for users who would be
concerned about the GPL dependency from their test-scripts.

A new control `HdmiCecControl` will be added to `control_gpl.py` in the
future which will depend on the GPLv2 licenced libcec.