HdmiCecControl: Implemented CEC support using libCEC
In combination with a USB-CEC adaptor this allows stb-tester to send
keypresses to devices over HDMI-CEC.  This is useful for testing devices
which don't have an IR input like the Sony PlayStation 4.

The URI is of the form:

    hdmi-cec:[<device>]:[<source>]:[<destination>]

If device is not specified then any available CEC adapter will be used.

Similar to the keymapping for Roku, a map is used to lookup the CEC
commands.  We raise `UnknownKeyError` if it's not one we've heard of.  The
list of keys is from http://www.cec-o-matic.com .  I've done my own
interpretation on how they should be mapped onto linux input keynames.

The tests are implemented by monkey-patching out the bits of libcec that
actually interact with the hardware.  I'm not a great fan of this style
of testing, but in the absence of hardware-in-the-loop testing this is
the best we can do.  At least some of this code will actually get run
through during `make check`.

The tests are particularly ugly because the SWIG generated libcec bindings
makes inspecting `cec_commmand` objects inconvenient so I've had to use
`ctypes` :(.

This is based on an initial prototype and research by Daniel Andersson
<d.andersson92@gmail.com> (@danielandersson).  See PR #378 for more
information.