Ubuntu packaging: Build package for Ubuntu 16.04
And drop Ubuntu 15.04, which reached end of life in February 2016.

debuild seems to have changed its default compression for the
".debian.tar.gz" tarball it creates, from gzip to xz. This was causing
`build-source-package.sh` to fail later on when it expected a file with
the ".gz" suffix.

Then, dpkg-buildpackage was failing with "gpg: signing failed: No
PINentry" when it tried to sign the ".changes" file. Using debuild
(which is just a higher level wrapper around dpkg-buildpackage, AFAICT)
doesn't have this problem.