Packaging tests: Install /usr/bin/time
...into the pristine container used to test the stb-tester packages.
`time` isn't a run-time dependency of stb-tester, but
`test_auto_selftest_caching` needs it.

Also while I'm here tidy up the installation path where we install the
stb-tester packages from: It used to be `tmp/stb-tester*.deb`, now it's
`/tmp/stb-tester*.deb` (with a leading slash). It worked before because
the command-line was run from the root of the container, but this is
more explicit & clear.