extra/pylint.sh: Silence invalid pylint PyGIWarning
On Ubuntu 16.04, pylint/astroid prints the following message whenever
it tries to import `gi.repository.Gst` (note 2 separate lines):

> /usr/lib/python2.7/dist-packages/astroid/brain/brain_gi.py:136: PyGIWarning: Gst was imported without specifying a version first. Use gi.require_version('Gst', '1.0') before import to ensure that the right version gets loaded.
>   __import__(m)

...even though the code in stb-tester already does
`gi.require_version("Gst", "1.0")`. The source of this warning is
when *pylint* tries to import `Gst` as it's analysing the stb-tester
module, not when stb-tester itself imports `Gst`. This seems to be a bug
in pylint, that only became visible with the newer version of GObject
Introspection in Ubuntu 16.04. I get this warning with astroid 1.2.1
installed from pip (1.2.1 is the version we pinned for testing on Travis
on Ubuntu 14.04) and with astroid 1.4.4 (the version that ships with
Ubuntu 16.04).