HdmiCecControl: Implemented CEC support using libCEC
Added support to control devices using CEC.
Similar to the keymapping for Roku, a map is used to lookup the CEC commands.

Added support for automatic detection of CEC adapter when "hdmi-cec:auto" is set in stbt.conf

Need to solve the CEC-addressing.
Now it is hardcoded to (source: Recording 1, destination: Playback 1) 14.
Ex "14:44:00". 1 is ID of the source (CEC adapter). 4 is the destination (device under test)
44:00 is the command.