gst_utils: Add `CapturedFrame` class
This is an ndarray with a timestamp attached.  We will return these from
our get_frame() calls, etc. to get the best of both worlds of
`numpy.ndarray` and `Gst.Sample`.

Subclassing `numpy.ndarray` is not entirely straightforward.  I used the
documentation from numpy here:
http://docs.scipy.org/doc/numpy-1.10.1/user/basics.subclassing.html#slightly-more-realistic-example-attribute-added-to-existing-array

I've had to add `CapturedFrame` to the list of pylint ignored classes
because its base-class is numpy.ndarray.  numpy is already in the list
because pylint can't see inside Python packages implemented in C.