stbt.MatchResult: frame and image are no longer optional
Constructing a MatchResult without `frame` and `image` has been
deprecated for over 2 years (since stb-tester 0.19). In recent commits
we've changed the constructor in a backwards-incompatible way
anyway (changing the first parameter from nanoseconds to seconds). We've
decided that the constructor isn't part of the public API, so there is
no point in maintaining this code.