stbt.match_text: Fix exception for single-channel images
The `__repr__` of `TextMatchResult` was raising an `IndexError` if
`match_text` was given a single-channel (black-and-white or greyscale)
image. This means that you couldn't do custom thresholding before
running the image through `match_text`. This was the stack trace:

    Traceback (most recent call last):
     [...]
     _stbt/core.py line 1036 in match_text
       debug("match_text: Match found: %s" % str(result))
     _stbt/core.py line 612 in __repr__
       self.frame.shape[2]),
    IndexError: tuple index out of range

Note that I have also changed `MatchResult.__repr__` for consistency,
but `match` probably needs other changes too before it can support
single-channel images.