stbt.ocr & stbt.match_text: Add `text_color` parameter
In low-contrast situations (for example white text on a blue background,
such as the "Summary" tab in `tests/action-panel.png`) tesseract's
binarisation algorithm often picks a threshold that removes the desired
text. So far we have worked around this in our test scripts by
specifying a region, so that the blue background is the only background
visible anywhere in the image, which usually helps tesseract to pick a
suitable threshold. But we can't always specify a region, for example
when the location of a menu varies depending on the contents of the
menu.

When you specify `text_color`, we perform binarisation ourselves based
on this information.

We use a simple euclidean distance in the BGR colorspace, which perhaps
isn't ideal but it seems to work. It might be worth testing the
weightings suggested in [1], or for something more computationally
expensive use CIE DeltaE-1994 [2]. I've picked an arbitrary threshold of
25, which is 10% of the possible range (0 - 255).

Note that we do the binarisation *after* scaling the image up, to
retain (some of) the information in the anti-aliasing of the text.

[1]: http://www.compuphase.com/cmetric.htm
[2]: http://stackoverflow.com/a/25726414/606705