stbt.match_text: Normalise unicode punctuation & ligatures
If tesseract reads the text "stb-tester" as u"stb\u2014tester" (\u2014
is the unicode code point for "em dash") then `stbt.ocr()` already
normalises it and returns "stb-tester" with an ascii dash.

But `stbt.match_text("stb-tester")` was returning False, because it
wasn't doing the same normalisation. Note that we have to normalise the
text that tesseract returns *and* the text that you give to match_text,
for backwards compatibility with users who previously worked around this
by saying `match_text(u"stb\u2014tester")`.