stbt auto-selftest: Fill the progress bar to 100%
Previously, the progress bar on the terminal would fill to one short of
the total:

    [################### ]       20 / 21 - Processing tests/xyz.py

Now the final state of the progress bar looks like this:

    [####################]  21 / 21

Note that you do get the "Processing tests/xyz.py" bit while it's
running, but not after it has finished.

I've also reduced the whitespace between the progress bar and the count.
I think it looks better. You're unlikely to have more than 1000 python
files in your test-pack, and even if you do it doesn't matter if the
numbers shift slightly when you start processing the 1000th file.