stbt lint: Compatibility with latest pylint/astroid version
Since astroid 1.4.0† (released November 2015, on the same day as pylint
1.5.0) they renamed `astroid.node_classes.{CallFunc => Call}`. Luckily
they added an alias for backwards compatibility; this alias is a proxy
class that implements `__instancecheck__` so `isinstance(..., CallFunc)`
still works.

I don't know why we were using `type(x) == X` instead of `isinstance`,
probably just an oversight.

Note that "the old names will be removed in astroid 2.0" so we'll have
to deal with that eventually.

† https://github.com/PyCQA/astroid/blob/astroid-1.4/ChangeLog#L101-L113