pylint.conf: Enable the "no-member" checker
We were disabling several checkers that no longer exist:

C0322, C0323 and C0324 used to be specific white-space cases. Now they
have been removed and for "backward compatibility" they are aliases for
C0326 (bad-whitespace) which is a much more general checker.

E1101 was "maybe-no-member"; now it is an alias for "no-member" which is
one of the most important checkers to have!

For all of the above this has been the case since at least pylint 1.3.1
which we upgraded to in 2a1470b two years ago.

R0921, R0922, R0923 and W0142 have been removed altogether (as of at
least pylint 1.6.4). In pylint 1.3.1 they were "Abstract class not
referenced", "Abstract class is only referenced %s times", "Interface
not implemented", and "Used * or ** magic".

I used the following script to find these cases, and I manually compared
the output of `pylint --help-msg` to see if the code it printed matched
the code I gave it:

    for code in $(
            sed -n '/^disable=/,/^$/ p' extra/pylint.conf |
            sed -e '/disable/d' -e 's/,//' -e 's/ //g'); do
        echo $code;
        pylint --help-msg=$code;
        read;
    done

Now that the "no-member" checker is enabled there were a few false
positives I had to suppress:

* Recent pylint versions required "--extension-pkg-whitelist" in order
  to load C extensions, as a security measure.
* We set `ignored-classes=generator` because, for context managers
  defined with `@contextmanager`, pylint doesn't know that they have
  `__enter__` and `__exit__` methods. "We don't understand decorators
  yet".[1]

[1]: https://github.com/PyCQA/pylint/issues/782#issuecomment-171425110