make check, .travis.yml: Upgrade to latest pylint
My main motivation is so that Travis checks that `stbt lint` works with
the current version of pylint. We were already installing a version that
is newer than Ubuntu 14.04's 1.1.0 (and older than Ubuntu 16.04's 1.5.2)
so we don't lose anything by upgrading (other than my entire morning --
sigh).

Pylint now has some checkers for PEP8-compliant import grouping, but it
doesn't check that in each group the imports are sorted, so we still
need to use `isort`.

This highlighted a bug in `test_pylint_plugin_on_itself` introduced
since c05809a two years ago; the older version of pylint didn't complain
if the file you gave to it didn't exist.