stbt.match_text: Add `case_sensitive` parameter
The current behaviour is to ignore case. With this parameter at least we
can disable this behaviour if we need to.

Note that none of the ligatures in `_ocr_transtab` (like ﬃ -> ffi) exist
in uppercase versions. There *are* other ligatures that do have
uppercase & lowercase versions (œ -> oe and Œ -> OE) but they aren't in
`_ocr_transtab` and adding them is outside the scope of this commit,
as I haven't seen them in the wild and finding some test cases would
take some time I don't have right now.