Add VNC RFB protocol remote control (#422)
The protocol is documented in [RFC6143].

This is used by Cisco to control some of their set-top boxes.  The protocol is
used, but there isn't actually a VNC server running on the devices.

Traditionally the RFB protocol uses key codes are the same as [used by X] (See
[RFC6143 section 7.5.4]), but this remote control implements Cisco-specific
keycodes.  If other keycodes are needed by somebody in the future we can
implement this then.

Co-authored by Fabrice Triboix <ftriboix@cisco.com>.

[RFC6143]: https://tools.ietf.org/html/rfc6143
[used by X]: https://cgit.freedesktop.org/xorg/proto/x11proto/plain/keysymdef.h
[RFC6143 section 7.5.4]: https://tools.ietf.org/html/rfc6143#section-7.5.4