stbt-batch: Performance improvements: Record screenshots in stbt-run
stbt batch run wants a JPG thumbnail of the system under test at the end
of a test run.  Previously it would spawn an additional python instance
after the test run, load the PNG screenshot and then save it as a jpeg.
If the test run succeeded it would even spawn yet another stbt run just
to create a screenshot.png before deleting it after converting it to JPG.

This commit tidies this up.  stbt-run is responsible for creating all
screenshots and thumbnails as it has access to the uncompressed source
screenshot to start with.  This can save 1s for each test run by
`stbt batch`.

There is a slight change in behaviour (it shouldn't affect real test
scripts, but it did affect the selftest
`test_that_stbt_run_can_print_exceptions_with_unicode_characters`): If
the test script fails without pulling any frames at all, we do save a
screenshot.png with the default setting of
`--save-screenshot=on-failure`.

Although the author of this commit is @wmanley, @drothlis has put just as
much work into this.