stbt-batch: Performance improvements: Record screenshots in stbt-run
stbt batch run wants a JPG thumbnail of the system under test at the end
of a test run.  Previously it would spawn an additional python instance
after the test run, load the PNG screenshot and then save it as a jpeg.
If the test run succeeded it would even spawn yet another stbt run just
to create a screenshot.png before deleting it after converting it to JPG.

This commit tidies this up.  stbt-run is responsible for creating all
screenshots and thumbnails as it has access to the uncompressed source
screenshot to start with.