Optimisation: Don't imprt stbt in stbt config
Importing `stbt` imports a whole bunch of other libraries.  This
can even take 1s.  This way we only import config.py which in-turn
only depends on the Python standard library.

This makes stbt batch run faster as it needs to look up a bunch of
config keys to work out which scripts to run.