Create a translation layer for OpenCV v2 and v3
On Fedora 25, the default OpenCV version is v3 (although the Python module
is still called `cv2`). There are a few incompatibilies with the current
code, such as changed variable names, function signatures, function
returns (see `findContours`) and the complete removal of the `cv`
sub-module.

The key advantaged to upgrading to OpenCV 3 are [2]:
- new algorithms and performance enhancements
- better Python support
- Python 3 compatibility

This commit adds a new module `opencv_shim` [1] which acts as an interface
for the modules to OpenCV so that they work with both v2 and v3.
Essentially, all modules should access OpenCV functions and constants via
`opencv_shim`, which provides pass-through to everything that is common
between v2 and v3.

[1] shim: https://en.wikipedia.org/wiki/Shim_(computing)
[2] For the full advertised list of improvements:
    http://opencv.org/opencv-3-0.html