stbt.ocr: Passing region=None means no region, not Region.ALL
This was added for backwards compatibility in v0.21 (two and a half
years ago) but since then it has printed a warning that this change
would be made eventually.

This is consistent with `stbt.match_text`. In fact my motivation for
making this change is that the difference in behaviour caused me some
confusion when I was writing a test for the recent `text_color` pull
request. It's also consistent with the output of `Region.intersect`
which means that it's safe to pass regions created from
`Region.intersect` into `stbt.ocr`.