stbt-control-relay: Work around Python buffering
I was seeing test failures because stbt control relay wasn't picking up
the `SEND_ONCE` commands the test was sending.  This is due to the
buffering internal to Python's file iteration which is supposed to iterate
over lines, but does additional bufferring.

I discovered that this was the issue using `strace`.

I don't understand why this bug exhibits when using socket passing and not
without.

See also the relevant part of the implementation of `file.next()` in
cpython:

https://github.com/python/cpython/blob/2.7/Objects/fileobject.c#L2233-L2237