stbt.MatchResult: Remove custom equality & hash
These were added in 8e5b52f ("stbt.MatchResult: Define __eq__, __ne__
and __hash__") with the intention of using `MatchResult` objects with
`wait_until`'s `stable_secs` parameter. However I've become convinced
that it's a bad idea to have objects with different public attributes to
be considered equal to each other.

Furthermore, the `image` attribute can be a string or a `numpy.ndarray`
(depending on what you gave to `stbt.match`). If it's a numpy array,
this `__eq__` will blow up with a ValueError because you can't use numpy
arrays in a boolean context: `a == b` gives an array of bools, and
treating *that* as a bool raises ValueError: "The truth value of an
array with more than one element is ambiguous. Use a.any() or a.all()"

Similarly these `MatchResult` objects would be hashable if `image` is
a string, but unhashable for numpy arrays.

This never made it into an official stb-tester release so removing it
now isn't a breaking change.