stbt.wait_until: Don't compare values if you haven't asked for stable_secs
The `stable_secs` parameter was added in 306330b ("wait_until: Add
`stable_secs` parameter"). It hasn't been released yet. That
implementation could potentially cause existing code to fail, if the
values returned by the `callable_` function aren't comparable (that is,
they have a `__ne__` that raises `NotImplementedError`).

With `stable_secs` we do have to do the comparison, obviously, but that
behaviour is opt-in (you have to specify `stable_secs`).