opencv_shim: warn about findContours only once
Previously, the "You are using OpenCV v2, ..." warning message was
printed once for every call to `stbt.match` because `findContours` is
called in its call stack. This is far to often and makes the test output
much harder to read, as well as more-or-less doubling the length of the
log.

This commit changes the message to use the built-in `warnings` library,
which by default will only print the message once. However, it annoyingly
prints the source line from where the warning originates, which is not
helpful here and is annoying. This can be disabled by preventing the `line`
argument from being pass in.