_stbt/core: Handle findContours change in OpenCV 3.2.0
In versions of OpenCV < 3.2.0, the `findContours` function would not
return any contour points that were on the edges of the image - it would
create a 1 pixel boundary.  This was worked around in `_match_template`
by shifting positions after `findContours`.

In OpenCV 3.2.0 [1] this was fixed so that `findContours` now returns
contours on the image edges.  The work around can then be removed for
this version to prevent negative coordinates being introduced.

[1] https://github.com/opencv/opencv/commit/56c133d