Add ability to interrupt other threads
As long as they are blocked on `stbt.frames()` or `time.sleep()`

We add the ability to interrupt a thread that we've previously spawned -
as long as they are blocked on `get_frame` or `time.sleep`.  This takes
care of tearing threads down once we're done with them without requiring
the user to add any complicated logic for this purpose.  This is the
invasive part that requires changes to `_stbt.core`

This same mechanism could be used to implement a timeout `contextmanager`.

Example:

    def look_for_logo():
        return wait_until(lambda: match("logo.png"))

    t = threading.Thread(target=look_for_logo)
    t.start()

    interrupt(t)
    t.join()