Add `@spawn` threading helpers
Makes it easier to write multi-threaded test-scripts where one is waiting
for more than one event at a time.  This has two significant parts:

We make it easy to start a thread and get a value back from it.  This
helps ensure that if one of our threads raises an exception it is
propagated back to the main thread.

Example:

    @spawn()
    def look_for_logo():
        return wait_until(lambda: match("logo.png"))

    @spawn()
    def motion():
        return wait_for_motion()

    with look_for_logo, motion:
        look_for_logo.join()