OpenCV 3 Compatibility
Debian testing now has OpenCV 3 so I need this change to make stbt
development convenient on my PC.  I'm not certain that this is complete,
but it makes the tests pass.

The trickiest thing was a change to behaviour in `cv2.findContours`.  Its
changed the way it handles edges. I'm not 100% certain that I've maintained
the same behaviour, but `_find_contour_boxes` seems to give the same values
on travis as on my PC.

Apparently [56c133d45] in OpenCV is the cause of the change in behaviour.

[56c133d45]: https://github.com/opencv/opencv/commit/56c133d459248d17165d77eb902a8049680bf896