Add tests for stbt.match debug html
These are simple characterization tests. Hopefully they won't cause too
much maintenance burden but we do need something like this because there
have been several unnoticed regressions in the stbt-debug output.

We can't check the contents of the generated PNGs because they aren't
byte-for-byte identical on Travis vs. my laptop -- due to different
GStreamer versions? OpenCV? At least we can test that the expected
filenames are created, and we do validate the contents of the index.html
files.