tests: Test that OCR user words/patterns actually have an effect
With the previous test image, recent versions of Tesseract read the text
correctly even without tesseract_user_patterns or tesseract_user_words.

I had to upgrade pytest to use `pytest.param`, which allows us to
specify that some of the parametrize settings are expected to fail.

The pytest.ini "xfail_strict = true" causes the test-run to fail if an
"xfail" test passes. Otherwise it passes with status "XPASS".