stbt.ocr: tesseract_user_{words,patterns} can be a single string
When you only want to specify a single pattern, it was a common mistake
to specify a string instead of a list with a single string. This used to
treat the string as a list of patterns of a single character each, which
is not the effect you wanted.

Note that stbt.match_text doesn't take either of these parameters so I
don't need to touch that.