_stbt.core._AnnotatedTemplate.friendly_name: Return relative path
This is the path given by the caller of `stbt.match` (et al), instead of
an absolute path. This makes the debug output of `stbt.match` etc more
readable. It's still unambiguous because the behaviour of
`stbt.load_image` is deterministic. If you're not sure which absolute
path the relative path is resolving to, you can use `stbt run -vv` to
see the extra debug from `_find_user_file`.

This affects:

- The debug output from stbt.detect_match: "Searching for
  {friendly_name}".
- The debug output from stbt.wait_for_match: "Matched {friendly_name}".
- MatchTimeout.expected parameter.
  - Also the str representation of MatchTimeout: "Didn't find match for
    '{expected}' within {timeout_secs}".
- The contents of the debug HTML generated by `stbt run -vv`.