stbt.ocr & stbt.match_text: Add upsample parameter to disable upsampling
Stb-tester upsamples the image before passing it to tesseract (since
v0.20 in 2014). For details see
https://stb-tester.com/blog/2014/04/14/improving-ocr-accuracy

With this parameter you can *disable* that behaviour. This is useful if
you want to do your own pre-processing before calling `stbt.ocr`. For
example this would have made it possible to implement `text_color` in
your test-pack without modifying stb-tester.