stbt.FrameObject: Use thread-local storage for `in_is_visible` counter
This fixes the following race condition (introduced in the previous
commit): If one thread calls a public property while `is_visible` is
still being run in another thread, the first thread gets the property
value instead of `None` (note that falsey frame objects are supposed to
return `None` for all public properties).

Without the fix, `test_that_is_visible_and_properties_arent_racy` fails
about 9/10 runs on my laptop -- good enough for a test that we can run
locally if we're working on the FrameObject implementation. Adding more
threads doesn't make it more reproducible -- it's always the 2nd or 3rd
thread that fails. I suppose that once the first thread finishes, the
value of the property is cached.