stbt.FrameObject: Evaluate is_visible eagerly in constructor
This eliminates the race condition between calling `is_visible` in one
thread vs. calling another property in another thread (because
`is_visible` is called in the constructor, so it has already been
memoized before you can call any other property). The advantage is that
we don't need thread-local storage for `__in_in_visible`.

The downside is that the body of your `is_visible` property can't use
other attributes that you set in your class's constructor (`__init__`)
-- unless you set them *before* calling the super-class's `__init__`,
which is a bit counter-intuitive.