debian: Install gdb (`stbt batch run` dependency)
`stbt batch run` uses gdb to generate a backtrace.log file if a core
file (core dump) is present after the test-run. This is useful because
it ensures that gdb is run against the same libraries that the core file
was generated from. `stbt batch` has had this functionality since day
one, but we've never specified gdb as a dependency in the debian
packages. Without gdb installed, backtrace.log says "gdb: command not
found" instead of the backtrace.