FrameObject: Fix AttributeError with multiple threads
This was broken when you created a FrameObject in one thread and then read
it in another.

The existing test didn't catch the issue because it was checking "all" the
results, rather than the expected results.  The test was wrong so "all" the
(nonexistent) results were fine.  Now the test is more explicit about what
it wants.

See 78478e02a1 and the discussion on #483 for the background behind the
thread-local.

Traceback:

    Traceback (most recent call last):
      ...
      File ".../stb-tester/_stbt/core.py", line 1727, in __repr__
        args = ", ".join(("%s=%r" % x) for x in self._iter_attrs())
      File ".../stb-tester/_stbt/core.py", line 1727, in <genexpr>
        args = ", ".join(("%s=%r" % x) for x in self._iter_attrs())
      File ".../stb-tester/_stbt/core.py", line 1731, in _iter_attrs
        if self:
      File ".../stb-tester/_stbt/core.py", line 1739, in __nonzero__
        return bool(self.is_visible)
      File ".../stb-tester/_stbt/core.py", line 1660, in inner
        self._FrameObject__local.in_is_visible += 1
    AttributeError: 'thread._local' object has no attribute 'in_is_visible'