interpress_delay: Refactor: Use time.time() instead of datetime.now()
It's a bit weird that we used the mockable `self._time` for `sleep` but
not for getting the current time. It turns out that we don't actually
mock out the `time` in unit tests for `press` (only for testing
`wait_until`). This change means that we *could* use a mock time in
unit-tests for `press`, if we wanted to.