Fix timing of stbt.press visualisation on the output video
The text would be drawn immediately on the video frame currently being
processed, not taking into account the latency in the video pipeline, so
it would appear a fraction of a second too early in the video.

The difference is quite obvious using videotestsrc + `--control=test` +
a script that does `sleep(2); stbt.press("snow")`. I tried to write an
automated test but it was too flaky.